Mobile-RE-Toolkit/scripts/Android/Geiger/requirements.txt
typer>=0.9.0
rich>=13.0.0
GitPython>=3.1.0
prompt-toolkit>=3.0.0
requests>=2.31.0